Add types for Span component (#20)
This adds the Span component to our typings. It was actually already in there for a while so I just needed to export the typings. Sadly I learned that there is no way to safely type the children of a React component in typescript since they always return JSX.Element and there is no way to refine what element a component returns. I did add some typing so at least as a reader you can see that text can only take strings, numbers, and spans, but it isn't actually enforced. You can see [this issue](https://github.com/microsoft/TypeScript/issues/21699) for more information on this. Surprisingly this is the first time when I actually actively prefer Flow, which makes it [rather easy to do this ](https://flow.org/en/docs/react/children/#toc-only-allowing-a-specific-element-type-as-children).